requirements.txt
python-docx
lxml
Pillow
customtkinter
pytest
selenium
webdriver-manager
